feat: unified connection with different backend
