+++ packages/connection/src/electron/index.ts
@@ -1,0 +1,1 @@
+export * from './common-channel-handler';
